+++ apps/core-lib-dev/src/app/custom-localization.element.ts
@@ -27,7 +27,7 @@ const extraTemplates = new Map([
 const { setLocale } = gdsInitLocalization({
   extraLocales,
   extraTemplates,
-})
+} as any)
 
 const html = htmlTemplateTagFactory((strs, ...values) =>
   strs.map((s, i) => s + values[i]).join('')
@@ -63,6 +63,23 @@ export class AppElement extends HTMLElement {
             <gds-option value="3">Option 3</gds-option>
             <gds-option value="4">Option 4</gds-option>
           </gds-dropdown>
+
+          <gds-datepicker
+            id="datepicker"
+            name="datepicker"
+            label="Pick a date"
+            dateformat="y/m/d"
+            showWeekNumbers
+            onchange="console.log('date changed', this.value)"
+          >
+            <span slot="sub-label"
+              >This text sits in a slot, so app can handle translation however
+              it likes</span
+            >
+            <span slot="message">Same here</span>
+          </gds-datepicker>
+
+          <gds-calendar></gds-calendar>
 
           <button type="submit">Submit</button>
           <button type="reset">Reset</button>