+++ libs/angular/src/index.ts
@@ -13,3 +13,5 @@ export * from './lib/slider'
 export * from './lib/in-page-wizard'
 export * from './lib/context-menu'
 export * from './lib/sortable-list'
+
+export * from './lib/shared/core-element'