+++ libs/core/src/primitives/listbox/option.styles.ts
@@ -1,13 +1,16 @@
 import { css } from 'lit'
 
 const style = css`
-  :host {
-    padding: 0.5 1rem;
-    cursor: pointer;
-  }
+  @layer base, reset, transitional-styles;
+  @layer base {
+    :host {
+      padding: 0.5 1rem;
+      cursor: pointer;
+    }
 
-  :host(:hover) {
-    background-color: grey;
+    :host(:hover) {
+      background-color: grey;
+    }
   }
 `