+++ libs/core/src/primitives/popover/popover.ts
@@ -1,4 +1,4 @@
-import { LitElement, html, unsafeCSS } from 'lit'
+import { HTMLTemplateResult, LitElement, html, unsafeCSS } from 'lit'
 import { property, state } from 'lit/decorators.js'
 import { classMap } from 'lit/directives/class-map.js'
 import { msg } from '@lit/localize'
@@ -11,6 +11,7 @@ import {
   Placement,
 } from '@floating-ui/dom'
 
+import { GdsElement } from '../../gds-element'
 import { watch, watchMediaQuery } from '../../utils/decorators'
 import { gdsCustomElement } from '../../utils/helpers/custom-element-scoping'
 import { TransitionalStyles } from '../../utils/helpers/transitional-styles'
@@ -18,6 +19,7 @@ import { TransitionalStyles } from '../../utils/helpers/transitional-styles'
 import { topLayerOverTransforms } from './topLayerOverTransforms.middleware'
 
 import styles from './popover.styles'
+import { reference } from '@popperjs/core'
 
 /**
  * @element gds-popover
@@ -33,7 +35,7 @@ import styles from './popover.styles'
  * @fires gds-ui-state - Fired when the popover is opened or closed
  */
 @gdsCustomElement('gds-popover')
-export class GdsPopover extends LitElement {
+export class GdsPopover extends GdsElement {
   static styles = unsafeCSS(styles)
 
   /**
@@ -61,6 +63,20 @@ export class GdsPopover extends LitElement {
   @property()
   placement: Placement = 'bottom-start'
 
+  /**
+   * A callback that returns the minimum width of the popover.
+   * By default, the popover minWidth will be as wide as the trigger element.
+   */
+  @property()
+  calcMinWidth = (referenceEl: HTMLElement) => `${referenceEl.offsetWidth}px`
+
+  /**
+   * A callback that returns the maximum width of the popover.
+   * By default, the popover maxWidth will be set to `auto` and will grow as needed.
+   */
+  @property()
+  calcMaxWidth = (_referenceEl: HTMLElement) => `auto`
+
   @state()
   private _trigger: HTMLElement | undefined = undefined
 
@@ -81,6 +97,10 @@ export class GdsPopover extends LitElement {
     this.#registerAutoPositioning()
   }
 
+  // Used for Transitional Styles in some legacy browsers
+  @state()
+  private _tStyles?: HTMLTemplateResult
+
   // A reference to the dialog element used to make the popover modal
   #dialogElementRef: Ref<HTMLDialogElement> = createRef()
 
@@ -98,7 +118,9 @@ export class GdsPopover extends LitElement {
     this.addEventListener('keydown', (e: KeyboardEvent) => {
       if (e.key === 'Escape') {
         this.open = false
-        e.stopImmediatePropagation()
+        this.#dispatchUiStateEvent()
+        e.stopPropagation()
+        e.preventDefault()
       }
     })
 
@@ -123,22 +145,23 @@ export class GdsPopover extends LitElement {
   }
 
   render() {
-    return html`<dialog
-      class="${classMap({ 'v-kb-visible': this._isVirtKbVisible })}"
-      ${ref(this.#dialogElementRef)}
-    >
-      <header>
-        <h2>${this.label}</h2>
-        <button
-          class="close"
-          @click=${this.#handleCloseButton}
-          aria-label="${msg('Close')}"
-        >
-          <i></i>
-        </button>
-      </header>
-      <slot></slot>
-    </dialog>`
+    return html`${this._tStyles}
+      <dialog
+        class="${classMap({ 'v-kb-visible': this._isVirtKbVisible })}"
+        ${ref(this.#dialogElementRef)}
+      >
+        <header>
+          <h2>${this.label}</h2>
+          <button
+            class="close"
+            @click=${this.#handleCloseButton}
+            aria-label="${msg('Close')}"
+          >
+            <i></i>
+          </button>
+        </header>
+        <slot></slot>
+      </dialog>`
   }
 
   @watch('open')
@@ -167,11 +190,13 @@ export class GdsPopover extends LitElement {
         )
       }
     })
+  }
 
+  #dispatchUiStateEvent = () => {
     this.dispatchEvent(
       new CustomEvent('gds-ui-state', {
         detail: { open: this.open },
-        bubbles: true,
+        bubbles: false,
         composed: false,
       })
     )
@@ -181,6 +206,7 @@ export class GdsPopover extends LitElement {
     e.stopPropagation()
     e.preventDefault()
     this.open = false
+    this.#dispatchUiStateEvent()
 
     // The timeout here is to work around a strange default behaviour in VoiceOver on iOS, where when you close
     // a dialog, the focus gets moved to the element that is visually closest to where the focus was in the
@@ -236,7 +262,8 @@ export class GdsPopover extends LitElement {
         Object.assign(floatingEl.style, {
           left: `${x}px`,
           top: `${y}px`,
-          minWidth: `${referenceEl.offsetWidth}px`,
+          minWidth: this.calcMinWidth(referenceEl),
+          maxWidth: this.calcMaxWidth(referenceEl),
         })
       )
     })
@@ -249,9 +276,11 @@ export class GdsPopover extends LitElement {
     if (e.key === 'ArrowDown') {
       e.preventDefault()
       this.open = true
+      this.#dispatchUiStateEvent()
     }
     if (e.key === 'Escape') {
       this.open = false
+      this.#dispatchUiStateEvent()
     }
   }
 
@@ -268,10 +297,12 @@ export class GdsPopover extends LitElement {
     })
   }
 
-  #clickOutsideListener = (e: MouseEvent) => {
+  #clickOutsideListener = (evt: Event) => {
+    const e = evt as PointerEvent
     const dialog = this.#dialogElementRef.value
+    const isNotEnterKey = e.clientX > 0 || e.clientY > 0
 
-    if (dialog && this.open) {
+    if (isNotEnterKey && dialog && this.open) {
       const rect = dialog.getBoundingClientRect()
 
       const isInDialog =
@@ -281,7 +312,9 @@ export class GdsPopover extends LitElement {
         e.clientX <= rect.left + rect.width
 
       if (!isInDialog) {
+        e.stopPropagation()
         this.open = false
+        this.#dispatchUiStateEvent()
       }
     }
   }